Emit browser compatible build.
