+++ src/runtimes/browser.js
@@ -1,0 +1,1 @@
+export { loadFromNode, loadFromJSON } from '../index.js';
