+++ tests/runtimes/browser.js
@@ -1,0 +1,2 @@
+// import tests from shared to kick them off
+import '../shared';
