+++ tests/runtimes/node.js
@@ -1,0 +1,8 @@
+// import tests from shared to kick them off
+import '../shared';
+
+describe('heimdalljs-graph', function() {
+  describe('loadFromFile', function() {
+    it('works');
+  });
+});
